feat(webchat): expand drag-and-drop upload to the whole chat area
