+++ dashboard/src/components/chat/ChatInput.vue
@@ -2,9 +2,6 @@
   <div
     class="input-area fade-in"
     :class="{ 'is-dark': isDark }"
-    @dragover.prevent="handleDragOver"
-    @dragleave.prevent="handleDragLeave"
-    @drop.prevent="handleDrop"
   >
     <div
       class="input-container"
@@ -24,15 +21,6 @@
         transition: 'min-height 0.2s ease, padding 0.2s ease',
       }"
     >
-      <!-- 拖拽上传遮罩 -->
-      <transition name="fade">
-        <div v-if="isDragging" class="drop-overlay">
-          <div class="drop-overlay-content">
-            <v-icon size="48" color="primary">mdi-cloud-upload</v-icon>
-            <span class="drop-text">{{ tm("input.dropToUpload") }}</span>
-          </div>
-        </div>
-      </transition>
       <!-- 引用预览区 -->
       <transition name="slideReply" @after-leave="handleReplyAfterLeave">
         <div class="reply-preview" v-if="props.replyTo && !isReplyClosing">
@@ -418,12 +406,10 @@ const providerModelMenuRef = ref<InstanceType<typeof ProviderModelMenu> | null>(
 );
 const providerSelectorAvailable = ref(true);
 const isReplyClosing = ref(false);
-const isDragging = ref(false);
 const isComposing = ref(false);
 const inputIsMultiline = ref(false);
 const lastCompositionEndAt = ref<number | null>(null);
 const longPasteThreshold = 10_000;
-let dragLeaveTimeout: number | null = null;
 
 // 命令提示相关状态
 const allCommands = ref<CommandItem[]>([]);
@@ -927,35 +913,6 @@ function handlePaste(e: ClipboardEvent) {
   emit("pasteImage", e);
 }
 
-function handleDragOver(e: DragEvent) {
-  // 清除之前的 leave timeout
-  if (dragLeaveTimeout) {
-    clearTimeout(dragLeaveTimeout);
-    dragLeaveTimeout = null;
-  }
-
-  // 检查是否有文件
-  if (e.dataTransfer?.types.includes("Files")) {
-    isDragging.value = true;
-  }
-}
-
-function handleDragLeave(e: DragEvent) {
-  // 使用 timeout 避免在子元素间移动时闪烁
-  dragLeaveTimeout = window.setTimeout(() => {
-    isDragging.value = false;
-  }, 50);
-}
-
-function handleDrop(e: DragEvent) {
-  isDragging.value = false;
-
-  const files = e.dataTransfer?.files;
-  if (files && files.length > 0) {
-    emit("fileSelect", files);
-  }
-}
-
 function triggerImageInput() {
   imageInputRef.value?.click();
 }
@@ -1316,47 +1273,6 @@ defineExpose({
 .input-area:not(.is-dark) .input-action-btn:disabled {
   background: #f2f5f3 !important;
   color: rgba(0, 0, 0, 0.18) !important;
-}
-
-/* 拖拽上传遮罩 */
-.drop-overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(var(--v-theme-primary), 0.12);
-  border: 2px dashed rgba(var(--v-theme-primary), 0.45);
-  border-radius: 24px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 100;
-  pointer-events: none;
-}
-
-.drop-overlay-content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-}
-
-.drop-text {
-  font-size: 16px;
-  font-weight: 500;
-  color: rgb(var(--v-theme-primary));
-}
-
-/* Fade transition for drop overlay */
-.fade-enter-active,
-.fade-leave-active {
-  transition: opacity 0.2s ease;
-}
-
-.fade-enter-from,
-.fade-leave-to {
-  opacity: 0;
 }
 
 .reply-preview {
